refactor(evm): clean up transaction state API
Rename PendingTx to ExecutedTx, split state modules, inline the state-output docs, and remove TxGas and DatabaseCommit.
